Game

Description

This module should contain your game code.

Documentation

pegs :: Int

The number of pegs in a code. You can change this value if you want to play games with more than four pegs. You should make sure that you reference this definition in your code instead of hard-coding the number 4 or writing code that assumes this value is 4.

type Symbol = Char

Symbols are represented by characters.

symbols :: [Symbol]

The available symbols in the game.

type Code = [Symbol]

A code is represented by a list of symbols.

type Score = (Int, Int)

Guesses are scored using coloured and white markers. The first component of the pair gives the number of coloured markers and the right component gives the number of white markers.

data Player

A player is either human or computer-controlled.

Constructors

Human

Computer

codemaker :: Player

The first codemaker in a play session.

firstGuess :: Code

The first guess the AI will make.

correctGuess :: Score -> Bool

correctGuess score determines whether score indicates that a guess was correct or not. For example:

>>> correctGuess (pegs, 0)
True

>>> correctGuess (3, 1)
False

validateCode :: Code -> Bool

validateCode code checks that the code entered by a human player is valid. In other words, it should have the length given by pegs and it should only contain valid symbols. For example (assuming default definitions of pegs and symbols):

>>> validateCode "aabb"
True

>>> validateCode "zzyy"
False

>>> validateCode "ab"
False

codes :: [Code]

codes is a list of all possible codes. For example, assuming default definitions of pegs and symbols:

>>> codes
["aaaa", "aaab", "aaac", "aaad", .., "fffe", "ffff"]

results :: [Score]

results is a list of all possible scores. Assuming default definitions of pegs:

>>> results
[ (0,0), (0,1), (0,2), (0,3), (0,4), (1,0), (1,1), (1,2), (1,3), (2,0)
, (2,1), (2,2), (3,0), (4,0)
]

score :: Code -> Code -> Score

score code guess scores guess against code. Symbols which are in the right place and of the right type score a coloured marker. Symbols which are of the right type but in the wrong place score a white marker.

>>> score "abcd" "aabb"
(1,1)

nextGuess :: [Code] -> Code

nextGuess remainingOptions chooses the next guess from remainingOptions. If there is only one option left, choose it. Otherwise, calculate the hit score for each code and choose the code with the largest hit score.

eliminate :: Score -> Code -> [Code] -> [Code]

eliminate lastScore guess remainingOptions returns a list of Codes in which all codes from remainingOptions which would result in a different score for guess if they were the code are removed. In other words, given the set of remaining possible codes, narrow it down to those which would produce the same score we got from the codemaker.